+++ pasqal_cloud/client.py
@@ -170,7 +170,7 @@ class Client:
 
     def _get_workload(self, workload_id: str) -> Dict[str, Any]:
         workload: Dict[str, Any] = self._request(
-            "GET", f"{self.endpoints.core}/api/v1/workloads/{workload_id}"
+            "GET", f"{self.endpoints.core}/api/v2/workloads/{workload_id}"
         )["data"]
         return workload