+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.6] - 2023-09-28
+
+### Added
+
+Added `result_link` field to `Workload` object
+
+### Changed
+
+- `get_workload` now targets v2 of workloads endpoints
+- `result` is built from `result_link` where results are downloaded from temp s3 link
 
 ## [0.3.5] - 2023-07-21